// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  FLEET_MANAGER: 'FLEET_MANAGER',
  DISPATCHER: 'DISPATCHER',
  SAFETY_OFFICER: 'SAFETY_OFFICER',
  FINANCIAL_ANALYST: 'FINANCIAL_ANALYST'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const VehicleStatus = {
  AVAILABLE: 'AVAILABLE',
  ON_TRIP: 'ON_TRIP',
  IN_SHOP: 'IN_SHOP',
  RETIRED: 'RETIRED'
} as const

export type VehicleStatus = (typeof VehicleStatus)[keyof typeof VehicleStatus]


export const VehicleType = {
  TRUCK: 'TRUCK',
  VAN: 'VAN',
  BUS: 'BUS',
  TRAILER: 'TRAILER'
} as const

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType]


export const DriverStatus = {
  AVAILABLE: 'AVAILABLE',
  ON_TRIP: 'ON_TRIP',
  OFF_DUTY: 'OFF_DUTY',
  SUSPENDED: 'SUSPENDED'
} as const

export type DriverStatus = (typeof DriverStatus)[keyof typeof DriverStatus]


export const TripStatus = {
  DRAFT: 'DRAFT',
  DISPATCHED: 'DISPATCHED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type TripStatus = (typeof TripStatus)[keyof typeof TripStatus]


export const MaintenanceStatus = {
  OPEN: 'OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  CLOSED: 'CLOSED',
  CANCELLED: 'CANCELLED'
} as const

export type MaintenanceStatus = (typeof MaintenanceStatus)[keyof typeof MaintenanceStatus]


export const ExpenseType = {
  TOLL: 'TOLL',
  PARKING: 'PARKING',
  REPAIR: 'REPAIR',
  OTHER: 'OTHER'
} as const

export type ExpenseType = (typeof ExpenseType)[keyof typeof ExpenseType]